+++ server/src/main/java/com/project/Glog/controller/ReplyController.java
@@ -3,7 +3,7 @@ package com.project.Glog.controller;
 import com.project.Glog.dto.request.reply.ReplyCreateRequest;
 import com.project.Glog.dto.request.reply.ReplyGetRequest;
 import com.project.Glog.dto.request.reply.ReplyUpdateRequest;
-import com.project.Glog.dto.responsee.reply.ReplyGetResponse;
+import com.project.Glog.dto.response.reply.ReplyGetResponse;
 import com.project.Glog.security.CurrentUser;
 import com.project.Glog.security.UserPrincipal;
 import com.project.Glog.service.ReplyService;
