+++ server/src/main/java/com/project/Glog/controller/TemporaryController.java
@@ -1,7 +1,7 @@
 package com.project.Glog.controller;
 
-import com.project.Glog.dto.responsee.post.PostBasicDto;
-import com.project.Glog.dto.responsee.post.PostTitleResponse;
+import com.project.Glog.dto.response.post.PostBasicDto;
+import com.project.Glog.dto.response.post.PostTitleResponse;
 import com.project.Glog.security.CurrentUser;
 import com.project.Glog.security.UserPrincipal;
 import com.project.Glog.service.TemporaryService;
@@ -9,6 +9,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
 
 @RestController
 public class TemporaryController {
@@ -33,11 +36,12 @@ public class TemporaryController {
     }
 
 
-    @PostMapping("/temporary")
+    @PostMapping("/temporary") //TODO 이미지 파일 업로드 필요
     public ResponseEntity<String> createTemporary(@CurrentUser UserPrincipal userPrincipal,
-                                                  @RequestBody PostBasicDto postBasicDto){
+                                                  @RequestParam(value = "thumbnail", required = false) MultipartFile multipartFile,
+                                                  @RequestPart PostBasicDto postBasicDto) throws IOException {
 
-        temporaryService.create(userPrincipal, postBasicDto);
+        temporaryService.create(userPrincipal, multipartFile, postBasicDto);
 
         return new ResponseEntity<>("success create temporary",HttpStatus.OK);
     }
