+++ server/src/main/java/com/project/Glog/controller/UserController.java
@@ -1,8 +1,7 @@
 package com.project.Glog.controller;
 
-import com.project.Glog.dto.request.user.UserCreateRequest;
 import com.project.Glog.dto.request.user.UserInfoChangeRequest;
-import com.project.Glog.dto.responsee.user.UserDetailResponse;
+import com.project.Glog.dto.response.user.UserDetailResponse;
 import com.project.Glog.security.CurrentUser;
 import com.project.Glog.security.CustomUserDetailsService;
 import com.project.Glog.security.UserPrincipal;
