+++ server/src/main/java/com/project/Glog/dto/response/post/PostPreviewDto.java
@@ -1,17 +1,13 @@
-package com.project.Glog.dto.responsee.post;
+package com.project.Glog.dto.response.post;
 
 import com.project.Glog.domain.Post;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.time.LocalDateTime;
-
 @AllArgsConstructor
-@NoArgsConstructor
 @Getter @Setter
-public class PostReadResponse {
+public class PostPreviewDto {
     private Long postId;
     private String title;
     private String content;
@@ -19,22 +15,16 @@ public class PostReadResponse {
     private Integer likesCount;
     private Integer viewsCount;
     private String hashtags;
-    private LocalDateTime createdAt;
 
-    public static PostReadResponse of(Post post) {
-        return new PostReadResponse(
-                post.getId(),
+    //TODO 어떤 정보 보여줄지 기억이 안남 API 명세 작성 필요.
+
+    public static PostPreviewDto of (Post post){
+        return new PostPreviewDto(post.getId(),
                 post.getTitle(),
                 post.getContent(),
                 post.getImageUrl(),
                 post.getLikesCount(),
                 post.getViewsCount(),
-                post.getHashtags(),
-                post.getCreatedAt());
+                post.getHashtags());
     }
-
 }
-
-
-
-
