+++ server/src/main/java/com/project/Glog/dto/response/post/PostPreviewDtos.java
@@ -1,4 +1,4 @@
-package com.project.Glog.dto.responsee.post;
+package com.project.Glog.dto.response.post;
 
 import com.project.Glog.domain.Post;
 import lombok.Getter;
