+++ server/src/main/java/com/project/Glog/service/BlogService.java
@@ -3,16 +3,12 @@ package com.project.Glog.service;
 import com.project.Glog.domain.Blog;
 import com.project.Glog.domain.User;
 import com.project.Glog.dto.request.user.UserCreateRequest;
-import com.project.Glog.dto.responsee.blog.MyPageResponse;
+import com.project.Glog.dto.response.blog.MyPageResponse;
 import com.project.Glog.repository.BlogRepository;
 import com.project.Glog.repository.UserRepository;
 import com.project.Glog.security.UserPrincipal;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
-import java.util.Optional;
 
 @Service
 public class BlogService {
