+++ server/src/main/java/com/project/Glog/service/PostService.java
@@ -5,9 +5,9 @@ import com.project.Glog.domain.PostLike;
 import com.project.Glog.domain.User;
 import com.project.Glog.dto.request.post.PostCreateRequest;
 import com.project.Glog.dto.request.post.PostUpdateRequest;
-import com.project.Glog.dto.responsee.post.PostPreviewDtos;
-import com.project.Glog.dto.responsee.post.PostPreviewResponse;
-import com.project.Glog.dto.responsee.post.PostReadResponse;
+import com.project.Glog.dto.response.post.PostPreviewDtos;
+import com.project.Glog.dto.response.post.PostPreviewResponse;
+import com.project.Glog.dto.response.post.PostReadResponse;
 import com.project.Glog.repository.*;
 import com.project.Glog.security.UserPrincipal;
 import com.project.Glog.util.AwsUtils;
